Add version to to_frame()
To support future protocol versions, it's necessary to be able to
specify the desired frame version to return when converting a
protocol message into a frame.  This adds the "version" argument
to heyu.protocol.Message.to_frame(), defaulting it to the current
version.  Note that the parameter is currently unused.