Create a HubAction
The argparse package allows executing arbitrary actions on command-line
arguments.  This change creates and uses a HubAction, which will parse
the hub specified on the command line and store an appropriate tuple in
the namespace.  This also adds a function to determine the default hub.
Finally, a minor bug was fixed in the declared send_notification():
send_notification() was expecting the hub in the "hub" argument, but
the command line option would have stored into the "host" argument.