Move _hub_app assignment
In the start() method, we call the Tendril manager's connect() method,
then set _hub_app to True.  However, connect() isn't asynchronous, so
the _acceptor() method will have assigned a value to _hub_app by the
time connect() returns, and thus we will overwrite it.  This moves the
assignment to just before initiating the connect.