Abandon HubAction for a type
The argument parser will pass a string to the callable specified using
the "type" keyword argument.  This allows us to use util.parse_hub()
directly in an argument specification, without needing to set a special
action.  This switches from the "action=util.HubAction" methodology to
the "type=util.parse_hub" methodology and removes the now-unneeded
HubAction class and its tests.