Rework script interpretation
The script notifier calls a command.  The previous incarnation expected
the script expression to be a single string, which would then be split
up using shlex.  This approach uses the shell to split the script
arguments, and uses "type=" in the argument specification to prevent
unknown substitutions from being specified.