Add the gtk notification driver
Note that this driver does not work if PyGTK is not installed.  The
PyGTK package cannot be installed via a normal "pip install", so we
can't list it as a dependency; instead, we use conditional import
and a fake init() function that raises an explanatory exception
when the package is not available.