Begin retuning heyu
Currently, heyu is tied to gevent and the tendril networking library,
which doesn't seem to be working at present.  This begins the work to
recast it in terms of asyncio, using the framer library.  The only
code change at present is to remove a redundant str() call, but most
of the support files have been updated to reflect recent changes in
my Python package methodology, including supporting testing on
multiple versions of Python.